prix min max, retard de 1 entrée

+++ src/TouteLesVentes.java
@@ -146,6 +146,6 @@ public class TouteLesVentes {
 
 
     public List<Vente> trieVenteIntervalle(String prixMin, String prixMax) throws SQLException, ParseException {
-        return venteBD.ventePrixMinMaxList(Integer.valueOf(prixMin), Integer.valueOf(prixMax));
+        return venteBD.ventePrixMinMaxList(Double.valueOf(prixMin), Double.valueOf(prixMax));
     }
 }
